PID: Move from large tables to split tables
- Improve handler of TOF parametrization
- Add dedicated PID directory for PID tasks
- Add TOF beta task with dedicated QA
- Move from TOF full table to TOF split tables
- Move from TPC full table to TPC split tables
- Add QA tasks for split PID tables
- The PID split task is the PID default task
- Add separate task for the TOF beta
