+++ Analysis/Tasks/PID/pidTPC_tiny.cxx
@@ -17,20 +17,26 @@
 
 // O2 includes
 #include "Framework/AnalysisTask.h"
-#include "Framework/AnalysisDataModel.h"
-#include "Framework/ASoAHelpers.h"
 #include "Framework/HistogramRegistry.h"
-#include "Framework/runDataProcessing.h"
 #include "ReconstructionDataFormats/Track.h"
 #include <CCDB/BasicCCDBManager.h>
 #include "AnalysisDataModel/PID/PIDResponse.h"
 #include "AnalysisDataModel/PID/PIDTPC.h"
+#include "AnalysisDataModel/TrackSelectionTables.h"
 
 using namespace o2;
 using namespace o2::framework;
 using namespace o2::pid;
 using namespace o2::framework::expressions;
 using namespace o2::track;
+
+void customize(std::vector<o2::framework::ConfigParamSpec>& workflowOptions)
+{
+  std::vector<ConfigParamSpec> options{{"add-qa", VariantType::Int, 0, {"Produce TPC PID QA histograms"}}};
+  std::swap(workflowOptions, options);
+}
+
+#include "Framework/runDataProcessing.h"
 
 struct pidTPCTaskTiny {
   using Trks = soa::Join<aod::Tracks, aod::TracksExtra>;
@@ -129,7 +135,114 @@ struct pidTPCTaskTiny {
   }
 };
 
+struct pidTPCTaskQA {
+  static constexpr int Np = 9;
+  static constexpr const char* pT[Np] = {"e", "#mu", "#pi", "K", "p", "d", "t", "^{3}He", "#alpha"};
+  static constexpr std::string_view hexpected[Np] = {"expected/El", "expected/Mu", "expected/Pi",
+                                                     "expected/Ka", "expected/Pr", "expected/De",
+                                                     "expected/Tr", "expected/He", "expected/Al"};
+  static constexpr std::string_view hexpected_diff[Np] = {"expected_diff/El", "expected_diff/Mu", "expected_diff/Pi",
+                                                          "expected_diff/Ka", "expected_diff/Pr", "expected_diff/De",
+                                                          "expected_diff/Tr", "expected_diff/He", "expected_diff/Al"};
+  static constexpr std::string_view hnsigma[Np] = {"nsigma/El", "nsigma/Mu", "nsigma/Pi",
+                                                   "nsigma/Ka", "nsigma/Pr", "nsigma/De",
+                                                   "nsigma/Tr", "nsigma/He", "nsigma/Al"};
+  HistogramRegistry histos{"Histos", {}, OutputObjHandlingPolicy::QAObject};
+
+  Configurable<int> nBinsP{"nBinsP", 400, "Number of bins for the momentum"};
+  Configurable<float> MinP{"MinP", 0, "Minimum momentum in range"};
+  Configurable<float> MaxP{"MaxP", 20, "Maximum momentum in range"};
+
+  template <typename T>
+  void makelogaxis(T h)
+  {
+    const int nbins = h->GetNbinsX();
+    double binp[nbins + 1];
+    double max = h->GetXaxis()->GetBinUpEdge(nbins);
+    double min = h->GetXaxis()->GetBinLowEdge(1);
+    if (min <= 0) {
+      min = 0.00001;
+    }
+    double lmin = TMath::Log10(min);
+    double ldelta = (TMath::Log10(max) - lmin) / ((double)nbins);
+    for (int i = 0; i < nbins; i++) {
+      binp[i] = TMath::Exp(TMath::Log(10) * (lmin + i * ldelta));
+    }
+    binp[nbins] = max + 1;
+    h->GetXaxis()->Set(nbins, binp);
+  }
+
+  template <uint8_t i>
+  void addParticleHistos()
+  {
+    // Exp signal
+    histos.add(hexpected[i].data(), Form(";#it{p} (GeV/#it{c});d#it{E}/d#it{x}_(%s)", pT[i]), kTH2F, {{nBinsP, MinP, MaxP}, {1000, 0, 1000}});
+    makelogaxis(histos.get<TH2>(HIST(hexpected[i])));
+
+    // Signal - Expected signal
+    histos.add(hexpected_diff[i].data(), Form(";#it{p} (GeV/#it{c});;d#it{E}/d#it{x} - d#it{E}/d#it{x}(%s)", pT[i]), kTH2F, {{nBinsP, MinP, MaxP}, {1000, -500, 500}});
+    makelogaxis(histos.get<TH2>(HIST(hexpected_diff[i])));
+
+    // NSigma
+    histos.add(hnsigma[i].data(), Form(";#it{p} (GeV/#it{c});N_{#sigma}^{TPC}(%s)", pT[i]), kTH2F, {{nBinsP, MinP, MaxP}, {200, -10, 10}});
+    makelogaxis(histos.get<TH2>(HIST(hnsigma[i])));
+  }
+
+  void init(o2::framework::InitContext&)
+  {
+    // Event properties
+    histos.add("event/vertexz", ";Vtx_{z} (cm);Entries", kTH1F, {{100, -20, 20}});
+    histos.add("event/tpcsignal", ";#it{p} (GeV/#it{c});TPC Signal", kTH2F, {{nBinsP, MinP, MaxP}, {1000, 0, 1000}});
+    makelogaxis(histos.get<TH2>(HIST("event/tpcsignal")));
+
+    addParticleHistos<0>();
+    addParticleHistos<1>();
+    addParticleHistos<2>();
+    addParticleHistos<3>();
+    addParticleHistos<4>();
+    addParticleHistos<5>();
+    addParticleHistos<6>();
+    addParticleHistos<7>();
+    addParticleHistos<8>();
+  }
+
+  template <uint8_t i, typename T>
+  void fillParticleHistos(const T& t, const float nsigma)
+  {
+    histos.fill(HIST(hnsigma[i]), t.p(), nsigma);
+  }
+
+  void process(aod::Collision const& collision, soa::Join<aod::Tracks, aod::TracksExtra,
+                                                          aod::pidRespTPCTEl, aod::pidRespTPCTMu, aod::pidRespTPCTPi,
+                                                          aod::pidRespTPCTKa, aod::pidRespTPCTPr, aod::pidRespTPCTDe,
+                                                          aod::pidRespTPCTTr, aod::pidRespTPCTHe, aod::pidRespTPCTAl,
+                                                          aod::TrackSelection> const& tracks)
+  {
+    histos.fill(HIST("event/vertexz"), collision.posZ());
+
+    for (auto t : tracks) {
+      // const float mom = t.p();
+      const float mom = t.tpcInnerParam();
+      histos.fill(HIST("event/tpcsignal"), mom, t.tpcSignal());
+      //
+      fillParticleHistos<0>(t, t.tpcNSigmaEl());
+      fillParticleHistos<1>(t, t.tpcNSigmaMu());
+      fillParticleHistos<2>(t, t.tpcNSigmaPi());
+      fillParticleHistos<3>(t, t.tpcNSigmaKa());
+      fillParticleHistos<4>(t, t.tpcNSigmaPr());
+      fillParticleHistos<5>(t, t.tpcNSigmaDe());
+      fillParticleHistos<6>(t, t.tpcNSigmaTr());
+      fillParticleHistos<7>(t, t.tpcNSigmaHe());
+      fillParticleHistos<8>(t, t.tpcNSigmaAl());
+    }
+  }
+};
+
 WorkflowSpec defineDataProcessing(ConfigContext const& cfgc)
 {
-  return WorkflowSpec{adaptAnalysisTask<pidTPCTaskTiny>(cfgc, TaskName{"pidTPC-tiny-task"})};
+  auto workflow = WorkflowSpec{adaptAnalysisTask<pidTPCTaskTiny>(cfgc, TaskName{"pidTPC-tiny-task"})};
+  if (cfgc.options().get<int>("add-qa")) {
+    workflow.push_back(adaptAnalysisTask<pidTPCTaskQA>(cfgc, TaskName{"pidTOFQA-task"}));
+  }
+  return workflow;
 }
