+++ Analysis/Tasks/SkimmingTutorials/spectraTPCProvider.cxx
@@ -88,7 +88,11 @@ struct TPCSpectraProviderTask {
   Filter collisionFilter = nabs(aod::collision::posZ) < vertexZCut;
   Filter trackFilter = (nabs(aod::track::eta) < trackEtaCut) && (aod::track::pt > trackPtCut) && (aod::track::isGlobalTrack == (uint8_t) true);
 
-  using TrackCandidates = soa::Filtered<soa::Join<aod::Tracks, aod::TracksExtra, aod::pidRespTPC, aod::TrackSelection>>;
+  using TrackCandidates = soa::Filtered<soa::Join<aod::Tracks, aod::TracksExtra,
+                                                  aod::pidRespTPCEl, aod::pidRespTPCMu, aod::pidRespTPCPi,
+                                                  aod::pidRespTPCKa, aod::pidRespTPCPr, aod::pidRespTPCDe,
+                                                  aod::pidRespTPCTr, aod::pidRespTPCHe, aod::pidRespTPCAl,
+                                                  aod::TrackSelection>>;
   void process(soa::Filtered<soa::Join<aod::Collisions, aod::EvSels>>::iterator const& collision, TrackCandidates const& tracks)
   {
     if (!collision.alias()[kINT7]) {
