+++ Analysis/Tasks/SkimmingTutorials/spectraTPCReference.cxx
@@ -85,7 +85,11 @@ struct TPCSpectraReferenceTask {
     histos.fill(HIST(hpt[i]), track.pt());
   }
 
-  using TrackCandidates = soa::Filtered<soa::Join<aod::Tracks, aod::TracksExtra, aod::pidRespTPC, aod::TrackSelection>>;
+  using TrackCandidates = soa::Filtered<soa::Join<aod::Tracks, aod::TracksExtra,
+                                                  aod::pidRespTPCEl, aod::pidRespTPCMu, aod::pidRespTPCPi,
+                                                  aod::pidRespTPCKa, aod::pidRespTPCPr, aod::pidRespTPCDe,
+                                                  aod::pidRespTPCTr, aod::pidRespTPCHe, aod::pidRespTPCAl,
+                                                  aod::TrackSelection>>;
   void process(soa::Filtered<aod::Collisions>::iterator const& collision, TrackCandidates const& tracks)
   {
     for (auto track : tracks) {
